+++ frontend/src/Style/styles.css
@@ -13,7 +13,7 @@
   height: 100%;
   background: rgba(0, 0, 0, 0.5); 
   z-index: -1;
-  backdrop-filter: blur(5px); 
+  filter: blur(5px);
 }
