+++ frontend/src/components/Emailvalidate/ValidError.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import Alert from '../components/Alert/Alert'; 
+import Alert from '../Alert/Alert'; 
 
 const EmailValidationFailurePage = () => {
     const navigate = useNavigate();
