+++ frontend/src/components/Home/About.jsx
