+++ frontend/tailwind.config.js
@@ -7,8 +7,8 @@ export default {
   theme: {
     extend: {
       backgroundImage: theme => ({
-        'login-bg': "url('./src/Images/Login-bgImg.jpg')",
-        'signup-bg': "url('./src/Images/Signup-bg-img.jpg')",
+        'login-bg': "url('./src/assets/Images/Login-bgImg.jpg')",
+        'signup-bg': "url('./src/assets/Images/signup-bgImg.jpg')",
       }),
      
     },
